vertex_ai/requirements.txt
# VulnHunter V4 Enhanced Training Pipeline Requirements

# Google Cloud and Vertex AI
google-cloud-aiplatform>=1.45.0
google-cloud-storage>=2.10.0
google-auth>=2.23.0

# Machine Learning Core
tensorflow>=2.15.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Data Processing
joblib>=1.3.0
scipy>=1.11.0

# Utilities
pathlib
datetime
json5
pyyaml>=6.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0

# Logging and Monitoring
google-cloud-logging>=3.8.0
tensorboard>=2.15.0

# Optional: Enhanced features
transformers>=4.30.0  # For advanced NLP if needed
torch>=2.0.0  # Alternative ML framework
plotly>=5.15.0  # For advanced visualizations
seaborn>=0.12.0  # Statistical plots